Done Page

+++ benselfies/views.py
@@ -4,6 +4,7 @@ import urllib
 from django.core.files.base import ContentFile
 from django.http import HttpResponse
 from django.shortcuts import render_to_response, redirect
+from django.utils.datastructures import MultiValueDictKeyError
 from django.views.decorators.csrf import csrf_exempt
 from benselfies.models import UserSubmission, UserImage
 
@@ -15,10 +16,18 @@ def home(request):
 
 
 def like(request):
+    try:
+        request.session["user"]
+    except KeyError as _:
+        return redirect(email)
     return render_to_response('like_page.html')
 
 
 def upload(request):
+    try:
+        request.session["user"]
+    except KeyError as _:
+        return redirect(email)
     return render_to_response('upload.html')
 
 
@@ -26,9 +35,9 @@ def upload(request):
 def add_media_file(request):
     #TODO: check and upload to server
     try:
-        submission = UserSubmission.objects.get(user_id=request.POST["user_id"])
+        submission = UserSubmission.objects.get(pk=request.session["user"])
     except UserSubmission.DoesNotExist:
-        submission = UserSubmission.objects.create(user_id=request.POST["user_id"])
+        return redirect(email)
     name = str(len(UserImage.objects.filter(submission=submission)))
     image = UserImage.objects.create(submission=submission)
     file_image = cStringIO.StringIO(urllib.urlopen(request.POST['image']).read())
@@ -41,9 +50,9 @@ def add_media_file(request):
 @csrf_exempt
 def add_custom_pic(request):
     try:
-        submission = UserSubmission.objects.get(user_id=request.POST["user_id"])
+        submission = UserSubmission.objects.get(pk=request.session["user"])
     except UserSubmission.DoesNotExist:
-        submission = UserSubmission.objects.create(user_id=request.POST["user_id"])
+        return redirect(email)
     name = str(len(UserImage.objects.filter(submission=submission)))
     image = UserImage.objects.create(submission=submission)
     file_content = ContentFile(str(request.POST['file']).split(',')[1].decode('base64'))
@@ -54,9 +63,13 @@ def add_custom_pic(request):
 
 def done(request, user_id):
     try:
-        submission = UserSubmission.objects.get(user_id=user_id)
+        request.session["user"]
+    except KeyError as _:
+        return redirect(email)
+    try:
+        submission = UserSubmission.objects.get(pk=request.session["user"])
     except UserSubmission.DoesNotExist:
-        return redirect(home)
+        return redirect(email)
     images = UserImage.objects.filter(submission=submission)
     images = [image.image for image in images]
     images = ["/media/" + image.name for image in images if "final-" in image.name]
@@ -64,22 +77,26 @@ def done(request, user_id):
 
 
 @csrf_exempt
-def email(request):
-    try:
-        submission = UserSubmission.objects.get(user_id=request.POST["user_id"])
-        submission.delete()
-    except UserSubmission.DoesNotExist:
-        pass
-    UserSubmission.objects.create(user_id=request.POST["user_id"], email=request.POST["email"])
-    return HttpResponse({}, content_type="application/json")
-
-@csrf_exempt
 def add_image(request):
     try:
-        submission = UserSubmission.objects.get(user_id=request.POST["user_id"])
+        submission = UserSubmission.objects.get(pk=request.session["user"])
     except UserSubmission.DoesNotExist:
-        submission = UserSubmission.objects.create(user_id=request.POST["user_id"])
+        return redirect(email)
     image = UserImage.objects.create(submission=submission)
     image.image = request.FILES['image']
     image.save()
     return HttpResponse(json.dumps({'url': "/media/" + image.image.name}), content_type="application/json")
+
+
+@csrf_exempt
+def email(request):
+    if request.method == "POST":
+        try:
+            submission = UserSubmission.objects.get(email=request.POST["email"])
+            submission.delete()
+        except UserSubmission.DoesNotExist:
+            pass
+        submission = UserSubmission.objects.create(email=request.POST["email"], user_name=request.POST["name"])
+        request.session["user"] = submission.id
+        return redirect(like)
+    return render_to_response('email.html')
